# api/docs/swagger.yaml
basePath: /
definitions:
  apidocs.ConfigData:
    properties:
      audio_codecs:
        items:
          $ref: '#/definitions/domain.AudioCodec'
        type: array
      global_config:
        $ref: '#/definitions/domain.GlobalConfig'
      hw_accels:
        items:
          $ref: '#/definitions/domain.HWAccel'
        type: array
      output_protocols:
        items:
          type: string
        type: array
      ports:
        $ref: '#/definitions/apidocs.ConfigPorts'
      stream_statuses:
        items:
          $ref: '#/definitions/domain.StreamStatus'
        type: array
      version:
        $ref: '#/definitions/version.Info'
      video_codecs:
        items:
          $ref: '#/definitions/domain.VideoCodec'
        type: array
      watermark_positions:
        items:
          $ref: '#/definitions/domain.WatermarkPosition'
        type: array
      watermark_types:
        items:
          $ref: '#/definitions/domain.WatermarkType'
        type: array
    type: object
...
  domain.GlobalConfig:
    properties:
      buffer:
        $ref: '#/definitions/config.BufferConfig'
      hooks:
        $ref: '#/definitions/config.HooksConfig'
      ingestor:
        $ref: '#/definitions/config.IngestorConfig'
      listeners:
        $ref: '#/definitions/config.ListenersConfig'
      log:
        $ref: '#/definitions/config.LogConfig'
      manager:
        $ref: '#/definitions/config.ManagerConfig'
      publisher:
        $ref: '#/definitions/config.PublisherConfig'
      server:
        $ref: '#/definitions/config.ServerConfig'
      sessions:
        $ref: '#/definitions/config.SessionsConfig'
      watermarks:
        $ref: '#/definitions/config.WatermarksConfig'
    type: object
...
info:
  contact: {}
  description: REST API for live stream configuration, ingest/publish lifecycle, DVR
    recordings, and event hooks.
  title: Open Streamer API
  version: "1.0"
paths:
  /config:
    get:
      description: Returns available hardware accelerators (OS-detected), static enum
        lists, publisher listener ports, and the current runtime GlobalConfig.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.ConfigData'
      summary: Get server configuration.
      tags:
      - system
    post:
      consumes:
      - application/json
      description: Merges the request body into the current GlobalConfig and applies
        it — services are started or stopped to match.
      parameters:
      - description: Partial configuration to merge
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.GlobalConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.ConfigUpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      summary: Partially update server configuration.
      tags:
      - system
...
  /config/transcoder/probe:
    post:
      consumes:
      - application/json
      description: P0 stub — always returns ok=true. Real capability probe will be
        reintroduced once the native libav pipeline lands.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.probeStubResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Probe transcoder capabilities (stub during native migration).
      tags:
      - system
  /config/yaml:
    get:
      description: Returns global_config, streams, and hooks bundled in a single YAML
        document. Pair with PUT /config/yaml to round-trip an editor.
      produces:
      - application/yaml
      responses:
        "200":
          description: YAML document
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get full system configuration as YAML
      tags:
      - system
    put:
      consumes:
      - application/yaml
      description: Replaces global_config, streams, and hooks with the request body.
        Validation errors are returned as a list. On success, the runtime manager
        + coordinator diff against the previous state and start/stop/restart services
        accordingly.
      parameters:
      - description: Full system configuration YAML document
        in: body
        name: body
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Replace full system configuration from YAML
      tags:
      - system
...
  /hooks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.HookList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List hooks
      tags:
      - hooks
    post:
      consumes:
      - application/json
      parameters:
      - description: Hook configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Hook'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.HookData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Create hook
      tags:
      - hooks
  /hooks/{hid}:
    delete:
      parameters:
      - description: Hook ID
        in: path
        name: hid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete hook
      tags:
      - hooks
    get:
      parameters:
      - description: Hook ID
        in: path
        name: hid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.HookData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get hook
      tags:
      - hooks
    put:
      consumes:
      - application/json
      parameters:
      - description: Hook ID
        in: path
        name: hid
        required: true
        type: string
      - description: Hook configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Hook'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.HookData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Update hook
      tags:
      - hooks
...
  /sessions:
    get:
      parameters:
      - description: Filter by stream code
        in: query
        name: stream
        type: string
      - description: Filter by protocol (hls|dash|rtmp|srt|rtsp)
        in: query
        name: proto
        type: string
      - description: Filter by status (active|closed)
        in: query
        name: status
        type: string
      - description: Max sessions to return (default 0 = no cap)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.SessionList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List play sessions
      tags:
      - sessions
  /sessions/{id}:
    delete:
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: kicked
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Kick a play session
      tags:
      - sessions
    get:
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PlaySession'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get a play session by ID
      tags:
      - sessions
  /streams:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List streams
      tags:
      - streams
  /streams/{code}:
    delete:
      parameters:
      - description: Stream code
        in: path
        name: code
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete stream
      tags:
      - streams
    get:
      parameters:
      - description: Stream code (a-zA-Z0-9_)
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get stream
      tags:
      - streams
    post:
      consumes:
      - application/json
      parameters:
      - description: Stream code
        in: path
        name: code
        required: true
        type: string
      - description: Partial or full stream document
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.StreamPutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamData'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.StreamData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Create or partial-update stream
      tags:
      - streams
...
  /templates:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List templates
      tags:
      - templates
  /templates/{code}:
    delete:
      parameters:
      - description: Template code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete template
      tags:
      - templates
    get:
      parameters:
      - description: Template code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get template
      tags:
      - templates
    post:
      consumes:
      - application/json
      parameters:
      - description: Template code
        in: path
        name: code
        required: true
        type: string
      - description: Template configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Template'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Create or update template
      tags:
      - templates
  /vod:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.VODMountList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List VOD mounts
      tags:
      - vod
    post:
      consumes:
      - application/json
      parameters:
      - description: VOD mount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.VODMount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.VODMountData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Create VOD mount
      tags:
      - vod
  /vod/{name}:
    delete:
      parameters:
      - description: Mount name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete VOD mount
      tags:
      - vod
    get:
      parameters:
      - description: Mount name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.VODMountData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get VOD mount
      tags:
      - vod
    put:
      consumes:
      - application/json
      parameters:
      - description: Mount name
        in: path
        name: name
        required: true
        type: string
      - description: VOD mount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.VODMount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.VODMountData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Update VOD mount
      tags:
      - vod
...
  /watermarks/{filename}:
    delete:
      parameters:
      - description: Asset filename
        in: path
        name: filename
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete a watermark asset
      tags:
      - watermarks
    get:
      parameters:
      - description: Asset filename (e.g. logo.png)
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.WatermarkAssetData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get watermark metadata
      tags:
      - watermarks
...
swagger: "2.0"